Fix bug in process_10x_with_seurat.R when using a secondary assay
The similarity check between the feature_reference.csv and the seurat
object would fail if the feautere was present multiple times, or if the
feature had a similar named feature in the RNA assay (due to how
make.unique is implemented in seurat). This will fix that with a
hand-wavey check that isn't too bad given the check itself is a formality.