Correct how USABLEMEMORY is calculated when using slurm, and add dry run
Previously, `USABLEMEMORY` was calculated too early leading to an
improper calculation based on default values. This PR modifies how
default global values are assigned, allowing user provided ones to
override the defaults, leading to a correct calculation.

Furthermore, this PR adds the functionality to echo the qsub command it
is about to run, and provides a `dry run` option that only prints the
command, but doesn't execute it.